Changed BatchConn's divisor from double to long
git-svn-id: http://svn.code.sf.net/p/petavision/code/sandbox@9621 7c0444ca-5550-0410-9008-faecdbe78eb5

Former-commit-id: c84009e9a19f18297f5c49562538a425cef00e92

+++ DepthLCA/src/BatchConn.cpp
@@ -137,7 +137,7 @@ int BatchConn::normalize_dW(int arbor_ID){
             int numpatchitems = nxp*nyp*nfp;
             pvwdata_t * dwpatchdata = get_dwDataHead(loop_arbor,kernelindex);
             for( int n=0; n<numpatchitems; n++ ) {
-               double divisor = numKernelActivations[loop_arbor][kernelindex][n];
+               long divisor = numKernelActivations[loop_arbor][kernelindex][n];
                for(int i = 0; i < clones.size(); i++){
                   divisor += clones[i]->getNumKernelActivations(loop_arbor, kernelindex, n);
                }
